App speed up (resource shrinking)
* halfway optimization

+++ app/src/main/java/org/timetable/schemester/ChatRoomModel.java
@@ -19,7 +19,7 @@ public class ChatRoomModel {
     void setMessage(String newMessage){ this.message = newMessage;}
     void setUser(String newUser){this.user = newUser;}
     void setTimeStamp(String newStamp){this.timeStamp = newStamp;}
-    static ArrayList<ChatRoomModel> setModel(String msg, String id, String stamp, int val){
+    static ArrayList<ChatRoomModel> setModel(String msg, String id, String stamp, long val){
         ArrayList<ChatRoomModel> models = new ArrayList<>();
         for(int i = 0; i<val;i++) {
             models.add(new ChatRoomModel(msg, id, stamp));
